requirements.txt
backoff==2.2.1
Flask==3.1.1
openai==1.97.0
openpyxl==3.1.5
pandas==2.3.1
pdf2image==1.17.0
Pillow==11.3.0
pytesseract==0.3.13
Requests==2.32.4
